requirements.txt
absl-py==0.8.1
astor==0.8.0
attrs==19.2.0
audioread==2.1.8
backcall==0.1.0
bleach==3.1.0
cffi==1.13.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
gast==0.2.2
google-pasta==0.1.8
grpcio==1.24.1
h5py==2.10.0
importlib-metadata==0.23
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.0
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
librosa==0.7.1
llvmlite==0.30.0
Markdown==3.1.1
MarkupSafe==1.1.1
mistune==0.8.4
mock==3.0.5
more-itertools==7.2.0
msgpack==0.6.2
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numba==0.46.0
numpy==1.17.2
opt-einsum==3.1.0
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.10.0
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
pynvim==0.3.2
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
resampy==0.2.2
scikit-learn==0.21.3
scipy==1.3.1
Send2Trash==1.5.0
six==1.12.0
SoundFile==0.10.2
tensorboard==1.12.2
tensorflow-estimator==1.14.0
tensorflow-gpu==1.12.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
torch==1.2.0
torchvision==0.4.0
tornado==6.0.3
traitlets==4.3.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
